Avoid removing modules added by Reactor from `sys.modules`.
Usually when multiple files try to import a module, the first time will
trigger an import, and the imported module will be cached in
`sys.modules`. The second time an import happens it will get the module
from `sys.module`.

It appears that in `__init__.py` we remove any new module that was added
by Reactor, apart from a specific white list.
The meaning of this is that if a module has a side effect when we import
it, if the same module will be imported it won't find it in
`sys.module`, and the side effect will happen again.

To be specific, it causes an issue with `torchao`, which its import triggers the following line:
```
lib.define("int_matmul(Tensor a, Tensor b) -> Tensor")
```

Which will throw this error when it's execute twice:
```
RuntimeError: Tried to register an operator (torchao::int_matmul(Tensor a, Tensor b) -> Tensor) with the same name and overload name multiple times.
```